+++ projects/amdsmi/amdsmi_cli/amdsmi_cli.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/python3
 # PYTHON_ARGCOMPLETE_OK
 #
 # Copyright (C) Advanced Micro Devices. All rights reserved.